sahim_project/requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
msgpack==1.1.0
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.0
redis==5.2.0
sqlparse==0.5.2
tzdata==2024.2
